include request id with error response
This fixes #19.

As an added benifit, neovim sends a "shutdown" message when closing,
waiting about 1 second until the language server responds before
terminating to give the server a chance to respond. Previously, this
caused a slight hiccup whenever neovim was closed.

By clearly stating that the "shutdown" message is not implemented by
using the "id" field, neovim now closes immediately.